requirements.txt
openai
fastapi
uvicorn
python-dotenv
requests
pytesseract
pillow
pydantic